requirements.txt
Django==1.11.5
djangorestframework==3.8.2
